[ffi_demo] cleanup/refactor, separated timg_serial out of ffi_demo, limited Python dependency to libarlffi
