+++ examples/ffi_demo/timg_starpu/env.sh
@@ -1,0 +1,2 @@
+export LD_LIBRARY_PATH=$HOME/.local/starpu/lib
+export PKG_CONFIG_PATH=$HOME/.local/starpu/lib/pkgconfig
